chore(move): lets move the examples
lets use a separate examples package and rename things to be more clear
