+++ kubernetes-tests/src/test/java/io/fabric8/openshift/client/examples/ListDeploymentConfigs.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.fabric8.openshift.client.server.mock;
+package io.fabric8.openshift.client.examples;
 
 import io.fabric8.kubernetes.client.KubernetesClientException;
 import io.fabric8.openshift.api.model.DeploymentConfig;
@@ -26,7 +26,7 @@ import java.util.List;
 
 import static org.junit.Assert.assertNotNull;
 
-public class DeploymentConfigs {
+public class ListDeploymentConfigs {
   public static void main(String[] args) {
     try {
       OpenShiftClient client = new DefaultOpenShiftClient();
